+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 Change Log
 ==========
+Version 1.4.3 *(2017-02-21)*
+----------------------------
+* New: Added set click listener method for title
+* New: Added `setSaveCurrentPosition` builder method to use the current position when switching mode
+* Fix: Fixed title update when only year changed
+* Fix: Fixed decorator rendering with custom tile dimensions on API Level 21
+* Fix: Added missing `dynamicHeightEnabled` save in `onSaveInstanceState()`
+* Fix: If we used `match_parent` for the MaterialCalendarView but the size available is not enough, the view was truncated
+
 Version 1.4.2 *(2016-10-10)*
 ----------------------------
 * New: match_parent is now supported by tileSize, tileWidth and tileHeight